# openapi.yaml
openapi: 3.0.3
info:
  title: NexaGate API
  version: 0.2.0
  description: Authenticated management and monitoring endpoints for NexaGate.
servers:
  - url: https://{host}
    variables:
      host:
        default: localhost
paths:
  /healthz:
    get:
      summary: Health check
      responses:
        '200': {description: Healthy}
  /api/metrics:
    get:
      summary: Get live system metrics
      security: [{sessionCookie: []}]
      responses:
        '200': {description: Metrics JSON}
        '401': {description: Authentication required}
  /metrics:
    get:
      summary: Prometheus metrics
      security: [{sessionCookie: []}]
      responses:
        '200':
          description: Prometheus text exposition
          content:
            text/plain: {}
  /ws/metrics:
    get:
      summary: Stream live metrics over WebSocket
      security: [{sessionCookie: []}]
      responses:
        '101': {description: WebSocket upgrade}
  /api/update-status:
    get:
      summary: Get panel update status
      security: [{sessionCookie: []}]
      responses:
        '200': {description: Update status}
  /users:
    get:
      summary: List users
      security: [{sessionCookie: []}]
      responses:
        '200': {description: HTML user management page}
  /api/users:
    get:
      summary: List users as JSON
      security: [{sessionCookie: []}]
      responses:
        '200': {description: User list}
    post:
      summary: Create a user
      security: [{sessionCookie: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: {type: string}
                days: {type: integer, minimum: 0}
      responses:
        '201': {description: User created}
  /api/audit:
    get:
      summary: Read recent audit events
      security: [{sessionCookie: []}]
      responses:
        '200': {description: JSON array of recent audit lines}
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: nexagate_session